Move corrector.h from src/ to interface/, update #includes where neeeded. Update pattern file cfg to use corrected jets.
