+++ L1Trigger/Phase2L1ParticleFlow/src/corrector.cc
@@ -1,4 +1,4 @@
-#include "L1Trigger/Phase2L1ParticleFlow/src/corrector.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/corrector.h"
 
 #include <iostream>
 #include <sstream>
